Move zne extrapolation module
